Revert "chore: split TLS detection logic"
This reverts commit 6c0780443ce7d29699df5b76b7d8d2acbcfd64ad.
