+++ linkerd/app/inbound/src/direct.rs
@@ -70,12 +70,15 @@ pub struct ClientInfo {
     pub local_addr: OrigDstAddr,
 }
 
-type TlsIo<I> = tls::server::Io<identity::ServerIo<tls::detect_sni::DetectIo<I>>, I>;
+type TlsIo<I> = tls::server::Io<identity::ServerIo<tls::server::DetectIo<I>>, I>;
 type FwdIo<I> = SensorIo<io::PrefixedIo<TlsIo<I>>>;
 pub type GatewayIo<I> = FwdIo<I>;
 
 #[derive(Clone)]
-struct TlsParams(identity::Server);
+struct TlsParams {
+    timeout: tls::server::Timeout,
+    identity: identity::Server,
+}
 
 impl<N> Inbound<N> {
     /// Builds a stack that handles connections that target the proxy's inbound port
@@ -217,9 +220,11 @@ impl<N> Inbound<N> {
                 })
                 .push(svc::ArcNewService::layer())
                 .push(tls::NewDetectTls::<identity::Server, _, _>::layer(
-                    TlsParams(identity),
+                    TlsParams {
+                        timeout: tls::server::Timeout(detect_timeout),
+                        identity,
+                    },
                 ))
-                .push(tls::NewDetectSNI::layer(detect_timeout.into()))
                 .arc_new_tcp()
         })
     }
@@ -448,10 +453,17 @@ impl From<RefusedNoHeader> for Error {
 
 // === TlsParams ===
 
+impl<T> ExtractParam<tls::server::Timeout, T> for TlsParams {
+    #[inline]
+    fn extract_param(&self, _: &T) -> tls::server::Timeout {
+        self.timeout
+    }
+}
+
 impl<T> ExtractParam<identity::Server, T> for TlsParams {
     #[inline]
     fn extract_param(&self, _: &T) -> identity::Server {
-        self.0.clone()
+        self.identity.clone()
     }
 }
